# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /github/repos:
    get:
      operationId: listGithubRepos
      tags: [github]
      summary: List repositories the connected GitHub account can access
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Optional substring filter on owner/name.
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GithubRepo" }
        "403":
          description: Forbidden
        "502":
          description: GitHub integration error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GithubError" }
  /github/branches:
    get:
      operationId: listGithubBranches
      tags: [github]
      summary: List branches for a repo
      parameters:
        - name: repo
          in: query
          required: true
          schema: { type: string }
          description: owner/repo
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GithubBranch" }
        "403":
          description: Forbidden
        "502":
          description: GitHub integration error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GithubError" }
  /github/pulls:
    get:
      operationId: listGithubPulls
      tags: [github]
      summary: List pull requests for a repo
      parameters:
        - name: repo
          in: query
          required: true
          schema: { type: string }
          description: owner/repo
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Pulls
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GithubPull" }
        "403":
          description: Forbidden
        "502":
          description: GitHub integration error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GithubError" }
  /engineering-tasks/{id}/refresh-pr-state:
    post:
      operationId: refreshEngineeringTaskPrState
      tags: [engineering]
      summary: Refetch the PR state from GitHub and persist it on the task
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EngineeringTask" }
        "400":
          description: Task has no PR URL or URL is not parseable
        "403":
          description: Forbidden
        "404":
          description: Not found
        "502":
          description: GitHub integration error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GithubError" }
  /admin/engineering-space:
    get:
      operationId: adminEngineeringSpace
      tags: [admin]
      summary: Admin — Paperclip-backed Engineering Space (requests, child tasks, agents, scheduler status)
      responses:
        "200":
          description: Engineering Space payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineeringSpace"
        "403":
          description: Forbidden
  /admin/engineering-space/refresh:
    post:
      operationId: adminEngineeringSpaceRefresh
      tags: [admin]
      summary: Admin — trigger an immediate Paperclip push + poll
      responses:
        "202":
          description: Refresh acknowledged. Body reports whether each loop was started or already busy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EngineeringSpaceRefreshAck"
        "403":
          description: Forbidden
  /admin/engineering-space/{id}/retry-push:
    post:
      operationId: adminRetryPaperclipPush
      tags: [admin]
      summary: Admin — clear failure markers on one feature request and immediately re-push it to Paperclip. The only way to re-attempt after a failed push (the scheduler never auto-retries).
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Push succeeded
          content:
            application/json:
              schema:
                type: object
                required: [ok, paperclipIssueId]
                properties:
                  ok: { type: boolean, enum: [true] }
                  paperclipIssueId: { type: string }
        "409":
          description: Row missing, not planned, or already successfully pushed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        "502":
          description: Paperclip returned an error or was unreachable
          content:
            application/json:
              schema:
                type: object
                required: [ok, error]
                properties:
                  ok: { type: boolean, enum: [false] }
                  error: { type: string }
        "403":
          description: Forbidden
  /admin/engineering-space/{id}/retry-notion-push:
    post:
      operationId: adminRetryNotionPush
      tags: [admin]
      summary: Admin — clear Notion failure markers on one human-routed feature request and immediately re-create its Notion page. The only way to re-attempt after a failed push (the scheduler never auto-retries).
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Push succeeded
          content:
            application/json:
              schema:
                type: object
                required: [ok, notionPageId]
                properties:
                  ok: { type: boolean, enum: [true] }
                  notionPageId: { type: string }
        "409":
          description: Row missing or already successfully pushed to Notion
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        "502":
          description: Notion returned an error or was unreachable
          content:
            application/json:
              schema:
                type: object
                required: [ok, error]
                properties:
                  ok: { type: boolean, enum: [false] }
                  error: { type: string }
        "403":
          description: Forbidden
...
components:
  schemas:
    EngineeringSpace:
      type: object
      required: [schedulerStatus, notionSchedulerStatus, agents, requests]
      properties:
        schedulerStatus:
          $ref: "#/components/schemas/PaperclipSchedulerStatus"
        notionSchedulerStatus:
          $ref: "#/components/schemas/PaperclipSchedulerStatus"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/PaperclipAgent"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/EngineeringSpaceRequest"
    PaperclipSchedulerStatus:
      type: object
      required: [configured, pushRunning, pollRunning]
      properties:
        configured:
          type: boolean
        pushRunning:
          type: boolean
        pollRunning:
          type: boolean
        lastPushAt:
          type: ["string", "null"]
          format: date-time
        lastPollAt:
          type: ["string", "null"]
          format: date-time
        lastPushError:
          type: ["string", "null"]
        lastPollError:
          type: ["string", "null"]
    PaperclipAgent:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: ["string", "null"] }
        title: { type: ["string", "null"] }
        icon: { type: ["string", "null"] }
        status: { type: ["string", "null"] }
    EngineeringSpaceChildTask:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string }
        identifier: { type: ["string", "null"] }
        title: { type: string }
        status: { type: string }
        assigneeAgentId: { type: ["string", "null"] }
    EngineeringSpaceRequest:
      type: object
      required: [id, title, status, priority, paperclipIssueId, children]
      properties:
        id: { type: integer }
        title: { type: string }
        summary: { type: string }
        status:
          type: string
          enum: [requested, ready_for_execution, planned, in_progress, deployed]
        priority:
          type: string
          enum: [low, medium, high]
        adminPriorityRank: { type: ["integer", "null"] }
        engineeringOwner:
          type: ["string", "null"]
          enum: [agent, human, null]
        paperclipIssueId: { type: ["string", "null"] }
        paperclipIdentifier: { type: ["string", "null"] }
        paperclipStatus: { type: ["string", "null"] }
        paperclipPriority: { type: ["string", "null"] }
        paperclipAssigneeAgentId: { type: ["string", "null"] }
        paperclipPushedAt:
          type: ["string", "null"]
          format: date-time
        paperclipLastSyncedAt:
          type: ["string", "null"]
          format: date-time
        paperclipPushError: { type: ["string", "null"] }
        notionPageId: { type: ["string", "null"] }
        notionUrl: { type: ["string", "null"] }
        notionStatus: { type: ["string", "null"] }
        notionAssignee: { type: ["string", "null"] }
        notionPushedAt:
          type: ["string", "null"]
          format: date-time
        notionLastSyncedAt:
          type: ["string", "null"]
          format: date-time
        notionPushError: { type: ["string", "null"] }
        children:
          type: array
          items:
            $ref: "#/components/schemas/EngineeringSpaceChildTask"
    EngineeringSpaceRefreshAck:
      type: object
      required: [push, poll, schedulerStatus]
      properties:
        push: { type: string, enum: [started, busy] }
        poll: { type: string, enum: [started, busy] }
        notion:
          type: object
          properties:
            push: { type: string, enum: [started, busy] }
            poll: { type: string, enum: [started, busy] }
        schedulerStatus:
          $ref: "#/components/schemas/PaperclipSchedulerStatus"
        notionSchedulerStatus:
          $ref: "#/components/schemas/PaperclipSchedulerStatus"
...
    EngineeringTask:
      type: object
      required:
        - id
        - featureRequestId
        - title
        - description
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        featureRequestId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [backlog, todo, in_progress, in_review, done, cancelled]
        assigneeUserId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        assigneeEmail:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        githubBranch:
          type: ["string", "null"]
        githubPrNumber:
          type: ["integer", "null"]
        githubPrUrl:
          type: ["string", "null"]
        githubPrState:
          type: ["string", "null"]
        githubPrStateUpdatedAt:
          type: ["string", "null"]
          format: date-time
        createdByUserId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GithubRepo:
      type: object
      required: [fullName, name, private, htmlUrl, defaultBranch]
      properties:
        fullName: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        private: { type: boolean }
        htmlUrl: { type: string }
        defaultBranch: { type: string }
        updatedAt: { type: ["string", "null"], format: date-time }
    GithubBranch:
      type: object
      required: [name, protected, sha]
      properties:
        name: { type: string }
        protected: { type: boolean }
        sha: { type: string }
    GithubPull:
      type: object
      required: [number, title, state, merged, htmlUrl, headRef, baseRef, updatedAt]
      properties:
        number: { type: integer }
        title: { type: string }
        state: { type: string, enum: [open, closed] }
        merged: { type: boolean }
        htmlUrl: { type: string }
        headRef: { type: string }
        baseRef: { type: string }
        updatedAt: { type: string, format: date-time }
    GithubError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        status: { type: ["integer", "null"] }